Feature: Implement copyable concrete pool
Made some few tweaks on the previous design and fixed some bugs of its
dependency. The linked list example also got updated to demonstrate
concrete pool.

Had to workaround GCC bug 95291 involving non-type template parameters
of class type. This workaround introduce major drawback against ideal
design and is expected to be withdrawed as soon as the bug is resolved.
The buggy version of GCC would not be supported afterwards.

This change is breaking.