Feature: Add 'dcool::container::List'
Can be configured to use as dynamic or static capacity array.
'dcool::container::StaticList' is now an alias of a pre-configured
'dcool::container::List', without additional customizable configuration
which might be added back in the future.

Some old bugs are fixed along with this implementation.

This change is breaking for 'dcool::container::StaticList' with
customized exception safety strategy.