Feature: Add stuffed list
List can now be configured to be stuffed. A stuffed list will not
maintain seperated length and capacity, which means all insertion and
erasures will cause reallocation.

It is recommended to enable stuffed only for lists that rarely resizes
and will benefit from saving the extra space for recording length.